+++ packages/types/src/json.ts
@@ -777,9 +777,13 @@ export interface CommerceSubscriptionItemJSON extends ClerkResourceJSON {
   status: CommerceSubscriptionStatus;
   created_at: number;
   period_start: number;
-  period_end: number;
+  /**
+   * Period end is `null` for subscription items that are on the free plan.
+   */
+  period_end: number | null;
   canceled_at: number | null;
   past_due_at: number | null;
+  is_free_trial: boolean;
 }
 
 /**
@@ -809,6 +813,7 @@ export interface CommerceSubscriptionJSON extends ClerkResourceJSON {
   updated_at: number | null;
   past_due_at: number | null;
   subscription_items: CommerceSubscriptionItemJSON[] | null;
+  eligible_for_free_trial?: boolean;
 }
 
 /**